app/src/main/java/com/example/final_assessment/viewmodels/LoginViewModel.kt
package com.example.final_assessment.viewmodels

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.example.final_assessment.models.login.LoginResponse
import com.example.final_assessment.repositories.LoginRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class LoginViewModel @Inject constructor(private val loginRepository: LoginRepository) :
    ViewModel() {

    private val _loginResult = MutableLiveData<Result<LoginResponse>>()
    val loginResult: LiveData<Result<LoginResponse>> = _loginResult

    private val _isLoading = MutableLiveData<Boolean>()
    val isLoading: LiveData<Boolean> = _isLoading

    fun login(username: String, password: String) {
        _isLoading.value = true
        viewModelScope.launch {
            try {
                val response = loginRepository.login(username, password)
                if (response.isSuccessful) {
                    response.body()?.let {
                        _loginResult.value = Result.success(it)
                    } ?: run {
                        _loginResult.value = Result.failure(Exception("Empty response"))
                    }
                } else {
                    _loginResult.value =
                        Result.failure(Exception("Login failed: ${response.code()}"))
                }
            } catch (e: Exception) {
                _loginResult.value = Result.failure(e)
            } finally {
                _isLoading.value = false
            }
        }
    }
}
